refactor: prepare for packaging
