+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "swimpeek/cmd/swimpeek"
+
+func main() {
+	swimpeek.SwimPeek()
+}
